Renamed point collection example.
